requirements.txt
ultralytics==8.4.30
mlflow==3.10.1
pyyaml